Add auto kill functionality. Tests, just starting rc can be done using one script
