+++ package.json
@@ -35,11 +35,11 @@
     "scripts": {
         "clean": "rimraf lib *.jar *.log",
         "compile": "tsc",
-        "test": "node scripts/download-and-run-rc.js && mocha 'test/**/*.js'",
-        "test:unit": "mocha 'test/unit/**/*.js'",
-        "test:integration": "node scripts/download-and-run-rc.js && mocha 'test/integration/**/*.js'",
+        "test": "node scripts/test-util.js 'alltests'",
+        "test:unit": "node scripts/test-util.js 'unit'",
+        "test:integration": "node scripts/test-util.js 'integration'",
         "validate-user-code": "tsc --build test/user_code/tsconfig.json",
-        "precoverage": "node scripts/download-and-run-rc.js",
+        "precoverage": "node scripts/test-util.js 'startrc'",
         "coverage": "rimraf coverage && nyc node_modules/mocha/bin/_mocha 'test/**/*.js' -- --reporter-options mochaFile=report.xml --reporter mocha-junit-reporter",
         "pregenerate-docs": "rimraf docs",
         "generate-docs": "typedoc --options typedoc.json",
